Respect NO_PROXY env var in proxy_info_from_url(). (#58)
This allows users to specify an override value for the noproxy list, in
addition to overriding the proxy url, in proxy_info_from_url(). If no
override is given, we should respect the values in the no_proxy/NO_PROXY
environment variables, if present. This behavior can be seen in other
tools like curl (see curl/curl#1140).

I would also tweak this for the python3 module, but its ProxyInfo class
doesn't even seem to have a bypass_hosts attribute.